+++ src/click/decorators.py
@@ -4,7 +4,6 @@ import inspect
 import typing as t
 from collections import abc
 from functools import update_wrapper
-from gettext import gettext as _
 
 from .core import Argument
 from .core import Command
@@ -13,6 +12,7 @@ from .core import Group
 from .core import Option
 from .core import Parameter
 from .globals import get_current_context
+from .locales import gettext as _
 from .utils import echo
 
 if t.TYPE_CHECKING: