+++ src/click/parser.py
@@ -27,13 +27,13 @@ from __future__ import annotations
 import collections.abc as cabc
 import typing as t
 from collections import deque
-from gettext import gettext as _
-from gettext import ngettext
 
 from .exceptions import BadArgumentUsage
 from .exceptions import BadOptionUsage
 from .exceptions import NoSuchOption
 from .exceptions import UsageError
+from .locales import gettext as _
+from .locales import ngettext
 
 if t.TYPE_CHECKING:
     from .core import Argument as CoreArgument